+++ tp1/src/sorting.h
@@ -1,7 +1,7 @@
 #ifndef _SORTING_H_
 #define _SORTING_H_
 
-void sortAttackList(char *addr, size_t filesize, size_t available_mem,
+void sortAttackList(char *addr, size_t filesize, int available_mem,
     int ways, char *r_prefix, char *w_prefix);
 
 #endif
